% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cutoff_filter}
\alias{cutoff_filter}
\title{Apply cutoff filter to RGB pixel}
\usage{
cutoff_filter(pixel, threshold = 127)
}
\arguments{
\item{pixel}{A raw vector of length 3 (RGB).}

\item{threshold}{Integer threshold for greyscale cutoff (default is 200).}
}
\value{
A raw RGB vector representing either black or magenta.
}
\description{
Converts a pixel to black or magenta depending on its greyscale value and a threshold.
}
\examples{
this function is the argument of the apply_filter_to_bitmap

}
