% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{display_filtered_image}
\alias{display_filtered_image}
\title{Display bitmap image using magick}
\usage{
display_filtered_image(filtered_image)
}
\arguments{
\item{filtered_image}{A 3D raw array of filtered pixel values.}
}
\value{
Displays the image using magick and returns nothing.
}
\description{
Displays a filtered image by converting the bitmap to a viewable raster format.
}
\examples{
display_filtered_image(grey_img)

}
